Move inference code out of sampling.rs into its own module
