+++ crates/cli/src/introspection/mod.rs
@@ -1,7 +1,9 @@
+pub mod inference;
 pub mod sampling;
 pub mod type_unification;
 pub mod validation_schema;
 
-pub use sampling::{sample_schema_from_db, type_from_bson};
+pub use inference::type_from_bson;
+pub use sampling::sample_schema_from_db;
 pub use validation_schema::get_metadata_from_validation_schema;
